removes old unnecessary split out code
