+++ src/split_out.rs
@@ -5,130 +5,6 @@ use super::die;
 use super::verify;
 use super::git_helpers3;
 use super::cli::MgtCommandSplit;
-use gitfilter::filter::FilterOptions;
-use std::io::sink;
-
-// iterate over both the include, and include_as
-// repofile variables, and generate an overall
-// include string that can be passed to
-// git-filter-repo
-pub fn generate_split_out_arg_include(repofile: &RepoFile) -> Vec<String> {
-    let include = if let Some(include) = &repofile.include {
-        include.clone()
-    } else {
-        vec![]
-    };
-    let include_as = if let Some(include_as) = &repofile.include_as {
-        include_as.clone()
-    } else {
-        vec![]
-    };
-
-    let mut out_vec = vec![];
-    for path in include {
-        out_vec.push("--path".to_string());
-        if path == " " {
-            out_vec.push("".into());
-        } else {
-            out_vec.push(path);
-        }
-    }
-    // include_as is more difficult because the indices matter
-    // for splitting out, the even indices are the local
-    // paths, so those are the ones we want to include
-    for path in include_as.iter().step_by(2) {
-        out_vec.push("--path".to_string());
-        if path == " " {
-            out_vec.push("".into())
-        } else {
-            out_vec.push(path.into());
-        }
-    }
-
-    out_vec
-}
-
-// iterate over the include_as variable, and generate a
-// string of args that can be passed to git-filter-repo
-pub fn generate_split_out_arg_include_as(repofile: &RepoFile) -> Vec<String> {
-    let include_as = if let Some(include_as) = &repofile.include_as {
-        include_as.clone()
-    } else {
-        return vec![];
-    };
-
-    // sources are the even indexed elements, dest are the odd
-    let sources = include_as.iter().skip(0).step_by(2);
-    let destinations = include_as.iter().skip(1).step_by(2);
-    assert_eq!(sources.len(), destinations.len());
-
-    let pairs = sources.zip(destinations);
-    // pairs is a vec of tuples: (src, dest)
-    // when mapping, x.0 is src, x.1 is dest
-    let mut out_vec = vec![];
-    for (src, dest) in pairs {
-        out_vec.push("--path-rename".to_string());
-
-        // if user provided a single space to indicate
-        // move to root, then git-filter-repo wants that
-        // as an emptry string
-        let use_src = if src == " " { "" } else { src };
-        let use_dest = if dest == " " { "" } else { dest };
-
-        out_vec.push(format!("{}:{}", use_src, use_dest));
-    }
-    out_vec
-}
-
-pub fn generate_split_out_arg_exclude(repofile: &RepoFile) -> Vec<String> {
-    let mut out_vec = vec![];
-    match &repofile.exclude {
-        None => return out_vec,
-        Some(v) => {
-            out_vec.push("--invert-paths".to_string());
-            for path in v {
-                out_vec.push("--path".to_string());
-                if path == " " {
-                    out_vec.push("".into());
-                } else {
-                    out_vec.push(path.clone());
-                }
-            }
-            out_vec
-        },
-    }
-}
-
-pub fn perform_gitfilter(
-    filter_rules: gitfilter::filter::FilterRules,
-    output_branch: String,
-    dry_run: bool,
-    verbose: bool,
-) {
-    let filter_options = FilterOptions {
-        stream: sink(),
-        branch: Some(output_branch),
-        default_include: false,
-        with_blobs: false,
-    };
-
-    if dry_run || verbose {
-        println!("Running with filter rules:\n{:#?}", filter_rules);
-    }
-    if dry_run { return; }
-
-    let res = gitfilter::filter::filter_with_rules_direct(
-        filter_options, filter_rules);
-    if let Err(e) = res {
-        die!("Failed to perform gitfilter: {}", e);
-    }
-
-    // remember, at the end of gitfilter, we have to revert the files that
-    // are currently staged:
-    if let Err(e) = git_helpers3::reset_stage() {
-        die!("Failed to reset git stage after filer: {}", e);
-    }
-}
 
 pub fn run_split_out(
     cmd: &mut MgtCommandSplit,
@@ -171,7 +47,6 @@ pub fn run_split_out_from_repo_file(
     validate_repo_file(&mut repo_file, &mut cmd.output_branch);
     core::go_to_repo_root();
     core::safe_to_proceed();
-    let arg_strings = generate_arg_strings(&repo_file, cmd.verbose);
     let filter_rules = generate_gitfilter_filterrules(&repo_file, cmd.verbose);
     core::make_and_checkout_output_branch(
         &cmd.output_branch,
@@ -188,7 +63,7 @@ pub fn run_split_out_from_repo_file(
         Some(o) => o.clone(),
         None => die!("Failed to find output branch"),
     };
-    perform_gitfilter(filter_rules, output_branch, cmd.dry_run, cmd.verbose);
+    core::perform_gitfilter(filter_rules, output_branch, cmd.dry_run, cmd.verbose);
 
     // for split out, rebase is a bit different because
     // we actually need to fetch the remote repo|branch that
@@ -304,41 +179,6 @@ fn generate_gitfilter_filterrules(
     filter_rules
 }
 
-fn generate_arg_strings(
-    repo_file: &RepoFile,
-    verbose: bool,
-) -> core::ArgStrings {
-    let include_arg_str = generate_split_out_arg_include(repo_file);
-    let include_as_arg_str = generate_split_out_arg_include_as(repo_file);
-    let exclude_arg_str = generate_split_out_arg_exclude(repo_file);
-
-    let log_p = if verbose { " #  " } else { "" };
-    if verbose {
-        println!("{}include_arg_str: {}", log_p, include_arg_str.join(" "));
-        println!("{}include_as_arg_str: {}", log_p, include_as_arg_str.join(" "));
-        println!("{}exclude_arg_str: {}", log_p, exclude_arg_str.join(" "));
-    }
-
-    let include = if include_arg_str.len() != 0 {
-        Some(include_arg_str)
-    } else { None };
-
-    let include_as = if include_as_arg_str.len() != 0 {
-        Some(include_as_arg_str)
-    } else { None };
-
-    let exclude = if exclude_arg_str.len() != 0 {
-        Some(exclude_arg_str)
-    } else { None };
-
-
-    core::ArgStrings {
-        include_as,
-        include,
-        exclude,
-    }
-}
-
 pub fn get_remote_branch_from_args(
     cmd: &MgtCommandSplit,
 ) -> Option<&String> {
@@ -363,96 +203,4 @@ pub fn get_remote_branch_from_args(
 
     // should never get here but whatever
     return None;
-}
-
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn should_generate_include_args_properly() {
-        let mut repofile = RepoFile::new();
-        repofile.include_as = Some(vec![
-            "abc".into(), "abc-remote".into(),
-            "xyz".into(), "xyz-remote".into(),
-        ]);
-        repofile.include = Some(vec!["123".into()]);
-        let filter_args = generate_split_out_arg_include(&repofile);
-        assert_eq!(filter_args.join(" "), "--path 123 --path abc --path xyz");
-    }
-
-    #[test]
-    fn should_generate_exclude_args_properly_for_one_exclude() {
-        let mut repofile = RepoFile::new();
-        repofile.exclude = Some(vec![
-            "one".into(),
-        ]);
-        let filter_args = generate_split_out_arg_exclude(&repofile);
-        assert_eq!(filter_args.join(" "), "--invert-paths --path one");
-    }
-
-    #[test]
-    fn should_generate_exclude_args_properly_for_multiple_exclude() {
-        let mut repofile = RepoFile::new();
-        repofile.exclude = Some(vec![
-            "one".into(), "two".into(), "three".into(),
-        ]);
-        let filter_args = generate_split_out_arg_exclude(&repofile);
-        assert_eq!(filter_args.join(" "), "--invert-paths --path one --path two --path three");
-    }
-
-    // not sure how to test this. I want to test if
-    // println was called with certain strings (because dry-run was set true
-    // that means it will println instead of running the command)
-    // but i dont think you can do that in rust
-    // #[test]
-    // fn should_not_run_filter_exclude_if_no_exclude_provided() {
-    //     let matches = ArgMatches::new();
-    //     let mut runner = Runner::new(&matches);
-    //     // ensure we dont actually run anything
-    //     runner.dry_run = true;
-    //     let repofile = RepoFile::new();
-    //     runner.repo_file = repofile;
-    //     runner = runner.generate_arg_strings();
-    //     let exclude_arg_str = runner.exclude_arg_str.clone().unwrap();
-    //     assert_eq!(exclude_arg_str, "".to_string());
-    // }
-
-    #[test]
-    fn should_generate_empty_vecs_for_generating_args_of_none() {
-        let mut repofile = RepoFile::new();
-        repofile.exclude = None;
-        repofile.include = None;
-        repofile.include_as = None;
-
-        let filter_exclude_args = generate_split_out_arg_exclude(&repofile);
-        assert_eq!(filter_exclude_args.len(), 0);
-
-        let filter_include_args = generate_split_out_arg_include(&repofile);
-        assert_eq!(filter_include_args.len(), 0);
-
-        let filter_include_as_args = generate_split_out_arg_include_as(&repofile);
-        assert_eq!(filter_include_as_args.len(), 0);
-    }
-
-    #[test]
-    fn should_generate_split_out_arg_include_as_properly() {
-        let mut repofile = RepoFile::new();
-        repofile.include_as = Some(vec![
-            "abc-src".into(), "abc-dest".into(),
-            "xyz-src".into(), "xyz-dest".into(),
-        ]);
-        let filter_args = generate_split_out_arg_include_as(&repofile);
-        assert_eq!(filter_args.join(" "), "--path-rename abc-src:abc-dest --path-rename xyz-src:xyz-dest");
-    }
-
-    // if include_as is None, it shouldnt fail, but rather
-    // just return an empty string
-    #[test]
-    fn gen_split_out_arg_include_as_should_not_fail_if_no_include_as() {
-        let repofile = RepoFile::new();
-        let filter_args = generate_split_out_arg_include_as(&repofile);
-        assert_eq!(filter_args.len(), 0);
-    }
 }
